+++ packages/timescale/src/hyper.ts
@@ -1,6 +1,0 @@
-import { hyperStyled } from "@macrostrat/hyper";
-import styles from "./main.module.scss";
-
-const h = hyperStyled(styles);
-
-export default h;