+++ packages/timescale/src/index.ts
@@ -1,6 +1,11 @@
 import { defaultIntervals } from "./intervals";
 import { TimescaleProvider, useTimescale } from "./provider";
-import { Interval, TimescaleOrientation, IncreaseDirection } from "./types";
+import {
+  Interval,
+  TimescaleOrientation,
+  IncreaseDirection,
+  TimescaleClickHandler,
+} from "./types";
 import {
   TimescaleBoxes,
   Cursor,
@@ -11,8 +16,8 @@ import { nestTimescale } from "./preprocess";
 import { AgeAxis, AgeAxisProps } from "./age-axis";
 import classNames from "classnames";
 import { ScaleContinuousNumeric } from "d3-scale";
-import { ReactNode, useMemo } from "react";
-import h from "./hyper";
+import { ReactNode, useCallback, useMemo, useRef } from "react";
+import h from "./main.module.sass";
 
 export * from "./intervals-api";
 export type { Interval } from "./types";
@@ -22,9 +27,15 @@ export {
   defaultIntervals as intervals,
 };
 
-type ClickHandler = (event: Event, interval: any) => void;
+interface TimescaleDisplayProps {
+  intervalStyle?: IntervalStyleBuilder;
+  labelProps?: LabelProps;
+  onClick?: TimescaleClickHandler;
+  style?: React.CSSProperties;
+  className: string;
+}
 
-export interface TimescaleProps {
+export interface TimescaleProps extends TimescaleDisplayProps {
   intervals?: Interval[];
   orientation?: TimescaleOrientation;
   increaseDirection?: IncreaseDirection;
@@ -36,32 +47,54 @@ export interface TimescaleProps {
   rootInterval?: number;
   /** Configuration for the axis */
   axisProps?: Partial<AgeAxisProps>;
-  labelProps?: LabelProps;
-  onClick?: ClickHandler;
   cursorPosition?: number | null;
   cursorComponent?: any;
-  intervalStyle?: IntervalStyleBuilder;
   scale?: ScaleContinuousNumeric<number, number>;
+  rotateLabels?: boolean;
 }
 
-function TimescaleContainer(props: {
-  onClick?: ClickHandler;
-  className: string;
-  children?: ReactNode;
-}) {
-  const { onClick: clickHandler, ...rest } = props;
-  const { scale, orientation } = useTimescale();
+function TimescaleContainer(
+  props: TimescaleDisplayProps & {
+    children?: ReactNode;
+  },
+) {
+  const {
+    onClick: clickHandler,
+    intervalStyle,
+    labelProps,
+    children,
+    ...rest
+  } = props;
+  const { scale, orientation, timescale } = useTimescale();
 
-  function onClick(evt: any) {
-    const bbox = evt.currentTarget.getBoundingClientRect();
-    const pos =
-      orientation == TimescaleOrientation.HORIZONTAL
-        ? evt.clientX - bbox.x
-        : evt.clientY - bbox.y;
-    clickHandler(evt, scale.invert(pos));
-  }
+  const ref = useRef<HTMLDivElement | null>(null);
 
-  return h("div.timescale.timescale-container", { onClick, ...rest });
+  const onClick = useCallback(
+    (evt: any, interval: Interval | undefined) => {
+      // Outer click handler
+      const bbox = ref.current?.getBoundingClientRect();
+      let age: number | undefined = undefined;
+      if (bbox != null && scale != null) {
+        const pos =
+          orientation == TimescaleOrientation.HORIZONTAL
+            ? evt.clientX - bbox.x
+            : evt.clientY - bbox.y;
+        age = scale.invert(pos);
+      }
+      clickHandler(evt, { age, interval });
+    },
+    [clickHandler, scale, orientation],
+  );
+
+  return h("div.timescale.timescale-container", { ref, onClick, ...rest }, [
+    h(TimescaleBoxes, {
+      interval: timescale,
+      intervalStyle,
+      onClick,
+      labelProps,
+    }),
+    children,
+  ]);
 }
 
 export function Timescale(props: TimescaleProps) {
@@ -81,6 +114,7 @@ export function Timescale(props: TimescaleProps) {
     ageRange,
     length: l,
     showAgeAxis = true,
+    absoluteAgeScale,
     levels,
     scale,
     rootInterval = 0,
@@ -90,17 +124,27 @@ export function Timescale(props: TimescaleProps) {
     onClick = null,
     intervalStyle,
     increaseDirection = IncreaseDirection.DOWN_LEFT,
+    rotateLabels = false,
     labelProps,
+    style,
+    className,
   } = props;
 
   const [parentMap, timescale] = useMemo(() => {
-    if (intervals.length == 0) {
+    if (intervals == null || intervals.length == 0) {
       return [null, null];
     }
     return nestTimescale(rootInterval, intervals);
   }, [rootInterval, intervals]);
 
-  const className = classNames(orientation, "increase-" + increaseDirection);
+  const _className = classNames(
+    className,
+    orientation,
+    "increase-" + increaseDirection,
+    {
+      "rotate-labels": rotateLabels,
+    },
+  );
 
   if (parentMap == null || timescale == null) {
     return null;
@@ -112,22 +156,21 @@ export function Timescale(props: TimescaleProps) {
       timescale,
       selectedInterval: null,
       parentMap,
-      ageRange: ageRange,
+      absoluteAgeScale,
+      ageRange,
       length: l,
       orientation,
       levels,
       scale,
       increaseDirection,
     },
-    h(TimescaleContainer, { className }, [
-      h(TimescaleBoxes, {
-        interval: timescale,
-        intervalStyle,
-        onClick,
-        labelProps,
-      }),
-      h.if(showAgeAxis)(AgeAxis, axisProps),
-      h.if(cursorPosition != null)(cursorComponent, { age: cursorPosition }),
-    ]),
+    h(
+      TimescaleContainer,
+      { style, className: _className, intervalStyle, labelProps, onClick },
+      [
+        h.if(showAgeAxis)(AgeAxis, axisProps),
+        h.if(cursorPosition != null)(cursorComponent, { age: cursorPosition }),
+      ],
+    ),
   );
 }