+++ packages/timescale/stories/main.stories.ts
@@ -13,11 +13,42 @@ export const Vertical = {
   },
 };
 
+export const VerticalWithRotatedLabels = {
+  args: {
+    orientation: TimescaleOrientation.VERTICAL,
+    levels: [0, 5],
+    rotateLabels: true,
+  },
+};
+
+export const VerticalWithRotatedLabelsConstrained = {
+  args: {
+    orientation: TimescaleOrientation.VERTICAL,
+    levels: [1, 5],
+    rotateLabels: true,
+    style: {
+      width: 200,
+    },
+  },
+};
+
 export const Horizontal = {
   args: {
     orientation: TimescaleOrientation.HORIZONTAL,
     levels: [0, 5],
     absoluteAgeScale: false,
+    onClick: (e, interval) => {
+      console.log("Clicked interval:", interval);
+    },
+  },
+};
+
+export const HorizontalWithRotatedLabels = {
+  args: {
+    orientation: TimescaleOrientation.HORIZONTAL,
+    levels: [0, 5],
+    absoluteAgeScale: false,
+    rotateLabels: true,
     onClick: (e, interval) => {
       console.log("Clicked interval:", interval);
     },
@@ -33,13 +64,22 @@ export const HorizontalAbsolute = {
   },
 };
 
-export const Condensed = {
+export const HorizontalAbsoluteCondensed = {
   args: {
     orientation: TimescaleOrientation.HORIZONTAL,
     levels: [2, 4],
     absoluteAgeScale: true,
     length: 800,
     ageRange: [1000, 0],
+  },
+};
+
+export const HorizontalAbsoluteSuperCondensed = {
+  args: {
+    orientation: TimescaleOrientation.HORIZONTAL,
+    levels: [0, 5],
+    absoluteAgeScale: true,
+    length: 1000,
   },
 };