+++ packages/timescale/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0]- 2026-02-07
+
+- Improve handling of arbitrary timescales from the Macrostrat API
+- Improve interval and age click handling, returning `{interval, age}` as data
+  instead of just `interval`
+- Improve styling logic internally, switch from SCSS to SASS, and streamline CSS
+  module imports
+- Add a new `rotateLabels` prop to control label rotation
+- Improve UI component stories
+
 ## [3.0.1] - 2026-01-29
 
 - Change layout of `package.json`